# openapi.yaml
openapi: 3.1.0
info:
  title: Rewardz API
  description: |
    OpenAPI 3.1 surface for the 10 endpoints introduced or changed by
    TODO-0018 Colosseum Rewardz League. Nine of the ten endpoints are
    served by the Fastify `api/` service under the `/v1` prefix; the
    tenth (`GET /rewardz/proof/:authority`) is served by the
    `mvp-keeper-bot` Rust service and is documented here so SDK +
    mobile + mini-app consumers can work from a single contract.

    # Scope

    This spec documents ONLY the surface changed by TODO-0018. It is
    not a complete description of every route in the API; pre-existing
    endpoints (points balance, auth, quests GET, admin, etc.) are
    intentionally omitted to keep review tractable. New or
    behaviour-changed endpoints live here with full request/response
    schemas.

    # Contract pinning

    The response field names and HTTP status codes in this spec are
    the contract the SDK (`sdk/packages/sdk`) and mobile app consume.
    Any field rename in the underlying route handler MUST land here in
    lockstep or the SDK round-trip tests break.

    # Keeper vs API

    - Base URL for nine endpoints: `https://api.rewardz.example` (path
      prefix `/v1`).
    - Base URL for the keeper-served proof endpoint:
      `https://keeper.rewardz.example` (no `/v1` prefix; the keeper is
      a separate process with its own Axum router).

  version: 0.1.0-todo18
  contact:
    name: Rewardz team
    url: https://github.com/anthropics/rewardz
servers:
...
paths:
...
  /v1/referrals/attribute:
    post:
      tags: [referrals]
      summary: First-wins attribution of a wallet to a referral code
      description: |
        Called by the mini-app when a wallet connects after a
        `?r=code` deep link. Attribution is first-wins: the first
        referral to attach to a wallet sticks, subsequent calls for
        the same wallet are no-ops and return the existing attribution.

        Intentionally unauthenticated — gating on a signed message
        would break the deep-link flow. Rate-limited to 20 req/minute
        per IP to blunt referral-code enumeration and wallet-pinning.

        Code matching is case-insensitive server-side.
      operationId: attributeReferral
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferralAttributeRequest"
      responses:
        "200":
          description: |
            Attribution outcome. `attributed=true` means this call won
            the race; `attributed=false` means the wallet was already
            attributed and the response echoes the existing binding.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralAttributeResponse"
        "400":
          description: Malformed body (bad wallet or code format)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Referral code not recognised
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate-limited (20 req/minute per IP)
          headers:
            Retry-After:
              schema:
                type: integer
                minimum: 0
              description: Seconds until the rate-limit window resets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitedError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ───────────────────────────────────────────────────────────────
  # 4. POST /v1/airdrop/signup
  # ───────────────────────────────────────────────────────────────
...
  /v1/points/award:
    post:
      tags: [points]
      summary: Award points to a wallet (capacity-gated, idempotent)
      description: |
        Writes a positive points event to the off-chain ledger.

        # TODO-0018 changes

        - **Capacity gating:** when the caller resolves to a specific
          protocol (via API key), the award debits
          `protocols.remaining_capacity`. If the protocol is at zero
          capacity, the call fails with 409 CAPACITY_EXHAUSTED and
          surfaces the protocol_id for operator triage.
        - **Idempotency:** `idempotency_key` in both the body and the
          `Idempotency-Key` header is now required. Replays return the
          original result with `duplicate=true` rather than writing
          again.

        # Header vs body idempotency

        Both `Idempotency-Key` (HTTP header) and `idempotency_key` (JSON
        body) are required. The `requireIdempotencyKey` middleware
        enforces the header; the handler's body validator enforces
        the body field.

        # Authentication

        Requires `X-API-Key` (the protocol's API key). Unresolved
        webhooks (no API-key-scoped protocol) skip the capacity debit.
      operationId: awardPoints
      security:
        - apiKeyAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 128
            pattern: "^[A-Za-z0-9_-]+$"
          description: |
            Client-generated idempotency token — enforced by
            `requireIdempotencyKey` middleware (api/src/middleware/idempotency.ts).
            Must be 8-128 chars from `[A-Za-z0-9_-]` (UUIDs, ULIDs, hex, and
            base58 all pass). The same value must also be sent in the JSON
            body as `idempotency_key`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AwardPointsRequest"
      responses:
        "200":
          description: Points awarded (or replay of prior award)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AwardPointsResponse"
        "400":
          description: Missing or invalid fields (wallet_address, amount, reason, idempotency_key)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Protocol has no remaining capacity for this debit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapacityExhaustedError"
        "429":
          description: Rate-limited
          headers:
            Retry-After:
              schema:
                type: integer
                minimum: 0
              description: Seconds until the rate-limit window resets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitedError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ───────────────────────────────────────────────────────────────
  # 9. POST /v1/protocols/{id}/quests  (EXTENDED: multi-step)
  # ───────────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: signed-message
      description: |
        Protocol-owner bearer token obtained from a signed-message
        auth flow. Enforced by `requireBearerAuth +
        requireProtocolOwner` middleware on protocol-scoped admin
        endpoints.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Protocol API key (prefix `rwz_`). Generated at
        POST /v1/protocols/register. Used by server-to-server
        integrations calling `/v1/points/award`.
    walletAuth:
      type: apiKey
      in: header
      name: X-Wallet-Signature
      description: |
        Signed-message wallet auth. The exact header layout is
        enforced by `requireWalletAuth` middleware and includes the
        base58 pubkey, signature, and a replay-protection nonce.
  schemas:
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Short error code (e.g. "Bad Request", "Not Found")
        message:
          type: string
          description: Human-readable diagnostic
      example:
        error: Bad Request
        message: "Invalid body: wallet must be a valid base58 Solana pubkey"
    CapacityExhaustedError:
      type: object
      required: [error, message, protocol_id]
      properties:
        error:
          type: string
          enum: [CAPACITY_EXHAUSTED]
        message:
          type: string
        protocol_id:
          type: string
          format: uuid
      example:
        error: CAPACITY_EXHAUSTED
        message: Protocol has no remaining points capacity
        protocol_id: 00000000-0000-0000-0000-000000000000
...
    RateLimitedError:
      type: object
      required: [error, message, retry_after]
      properties:
        error:
          type: string
          enum: ["Too Many Requests"]
        message:
          type: string
        retry_after:
          type: integer
          minimum: 0
          description: |
            Seconds until the rate-limit window resets. Matches the
            `Retry-After` HTTP header emitted alongside the body.
      example:
        error: "Too Many Requests"
        message: "Rate limit exceeded. Try again in 12s."
        retry_after: 12
...
    Base58Pubkey:
      type: string
      pattern: "^[1-9A-HJ-NP-Za-km-z]{32,44}$"
      description: Base58-encoded Solana public key
    BigIntString:
      type: string
      pattern: '^-?\d+$'
      description: |
        BIGINT value serialised as a string to avoid JS `Number`
        precision loss. Clients may `Number()` when magnitude is
        known-safe.

    # ── League Join ───────────────────────────────────────────
...
    ReferralAttributeRequest:
      type: object
      required: [wallet, code]
      properties:
        wallet:
          $ref: "#/components/schemas/Base58Pubkey"
        code:
          type: string
          pattern: "^[A-Za-z0-9]{4,16}$"
          description: |
            Referral code. Case-insensitive (uppercased server-side
            before lookup). Pattern is deliberately permissive to
            accept both current 6-char Crockford codes emitted by
            league/join (`^[2-9A-HJ-NP-Z]{6}$`) and any legacy
            protocol-owned codes that predated the league allocator.
    ReferralAttributeResponse:
      type: object
      required: [attributed, protocolId, referralCode]
      properties:
        attributed:
          type: boolean
          description: true = this call won the first-wins race; false = wallet was already bound
        protocolId:
          type: string
          format: uuid
        referralCode:
          type: string

    # ── Airdrop ───────────────────────────────────────────────
...
    AwardPointsRequest:
      type: object
      required: [wallet_address, amount, reason, idempotency_key]
      properties:
        wallet_address:
          $ref: "#/components/schemas/Base58Pubkey"
        amount:
          type: integer
          minimum: 1
          description: Points to award (positive integer)
        reason:
          type: string
          description: Free-text reason (e.g. "quest_complete:abc123")
        idempotency_key:
          type: string
          minLength: 8
          maxLength: 128
          pattern: "^[A-Za-z0-9_-]+$"
          description: |
            Caller-generated idempotency key. MUST also be sent as
            the `Idempotency-Key` HTTP header with the SAME value; the
            middleware and body validator both enforce presence, and
            `point_events.source_reference` has a UNIQUE constraint that
            drives dedupe.
    AwardPointsResponse:
      type: object
      required: [success, duplicate]
      properties:
        success:
          type: boolean
        event_id:
          type: string
          description: |
            Primary key of the point_events row. OMITTED (JSON field
            absent, not null) when the service returns no event id —
            e.g. a duplicate replay that resolves without writing a
            new row. Clients should treat "field missing" and
            "duplicate=true" as equivalent signals.
        new_balance:
          $ref: "#/components/schemas/BigIntString"
          description: |
            Wallet's total points after the award (serialised as string
            to preserve BIGINT precision). OMITTED when the service
            returns no balance alongside the response (e.g. duplicate
            replays where the balance isn't re-queried).
        duplicate:
          type: boolean
          description: true = replay of a prior award with the same idempotency_key

    # ── Quests ────────────────────────────────────────────────
...